pass the OCTOEXTENSION version via the environment variables
(cherry picked from commit dbd6e5d29af815e5946248d82fd2ddc8f2f55133)

# Conflicts:
#	src/main/java/com/octopus/bamboo/plugins/task/pack/PackTask.java
