+++ src/main/java/com/octopus/bamboo/plugins/task/pack/PackTask.java.orig
@@ -19,6 +19,10 @@ import org.springframework.stereotype.Component;
 
 import javax.inject.Inject;
 import javax.inject.Named;
+<<<<<<< HEAD
+import java.io.File;
+=======
+>>>>>>> dbd6e5d... pass the OCTOEXTENSION version via the environment variables
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -118,6 +122,38 @@ public class PackTask extends OctoTask implements CommonTaskType {
             commands.addAll(Arrays.asList(myArgs));
         }
 
+<<<<<<< HEAD
+        final String cliPath = capabilityContext.getCapabilityValue(
+                OctoConstants.OCTOPUS_CLI_CAPABILITY + "." + octopusCli);
+
+        if (StringUtils.isNotBlank(cliPath) && new File(cliPath).exists()) {
+            commands.add(0, cliPath);
+
+            final ExternalProcess process = processService.executeExternalProcess(taskContext,
+                    new ExternalProcessBuilder()
+                            .command(commands)
+                            .workingDirectory(taskContext.getWorkingDirectory()));
+
+            return TaskResultBuilder.newBuilder(taskContext)
+                    .checkReturnCode(process, 0)
+                    .build();
+        }
+
+        commonTaskService.logError(
+                taskContext,
+                "OCTOPUS-BAMBOO-INPUT-ERROR-0003: The path of \"" + cliPath + "\" for the selected Octopus CLI does not exist.");
+        return TaskResultBuilder.newBuilder(taskContext).failed().build();
+=======
+        if (!StringUtils.isEmpty(commentParser)) {
+            try {
+                final CommentWorkItemHandler commentHandler = new CommentWorkItemHandler();
+                commentHandler.processComments((TaskContext) taskContext, commentParser, basePath);
+            } catch (Exception ex) {
+                return TaskResultBuilder.createFailedWithErrorResult(taskContext);
+            }
+        }
+
         return LaunchOcto(taskContext, commands);
+>>>>>>> dbd6e5d... pass the OCTOEXTENSION version via the environment variables
     }
 }
